fix(parse): extract shared mode parser to fix three drifts between tracker and opencode plugin
The opencode plugin mode parsing had drifted from the Claude Code tracker.
Extracted a shared parseModeChange() into src/hooks/caveman-parse.js used
by both consumers, fixing three confirmed divergences:

1. Brevity triggers missing in plugin — "less tokens" / "fewer tokens" /
   "be brief" / "be terse" / "shorter answers" now activate caveman in
   both paths (promised in README, worked only in tracker before).
2. Bogus level overwrites flag — expanded-template branch fell through
   to getDefaultMode() for unknown levels; now returns null so unknown
   levels leave the flag untouched, matching tracker behavior.
3. Independent modes unreachable — plugin expanded-template branch does
   not match /caveman-commit etc., so the flag never flipped and per-turn
   reinforcement kept injecting. Slash-command path still works correctly.

bin/install.js updated to copy caveman-parse.{js->cjs} alongside
caveman-config.{js->cjs} for the opencode plugin install path.

Closes #602
